remove old image form
